fix: メンテナンス画面からHome画面に戻る処理ができなくなってしまったため、一度戻しました
