+++ apps/app/lib/router/navigator/home_navigator.dart
@@ -1,7 +1,6 @@
 import 'package:cores_navigation/navigators.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_app/router/routes/home_route.dart';
-import 'package:flutter_app/router/routes/main_shell_route.dart';
 
 final class HomeNavigatorImpl implements HomeNavigator {
   const HomeNavigatorImpl();
