+++ apps/app/lib/router/navigator/setting_navigator.dart
@@ -1,6 +1,6 @@
 import 'package:cores_navigation/navigators.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_app/router/routes/main_shell_route.dart';
+import 'package:flutter_app/router/routes/home_route.dart';
 import 'package:flutter_app/router/routes/setting_route.dart';
 
 final class SettingNavigatorImpl implements SettingNavigator {
